Amjad50/OS: implemented `try_exist`, `flush`, `remove_dir_all` using sys_common
Not a big change at all, I was just looking at the functions marked as `todo` and these have basic implementation as such in `unix`, so I thought I'll use them.

We don't have `remove_dir_all` yet as it depends on `rmdir`, but it will be supported automatically once that is added.

For others, are basic functionalities.