Emerald: Fix bug when passing `file` as stdio
This is not good, we should replace with `dups` or other method to duplicate the file.

The issue was, when passing a file, we pass an `fd`, but we still own the file inside `Stdio::Fd`, which will be closed by the parent program, but its already moved to the new process.

this causes a crash on close